pyproject.toml
[tool.poetry]
name = "flask-sqlalchemy-testing"
version = "0.1.0"
description = ""
authors = ["Andy Jandl <akjandl@gmail.com>"]
readme = "README.md"
packages = [{include = "flask_sqlalchemy_testing"}]

[tool.poetry.dependencies]
python = "^3.9"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.9.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
mixer = "^7.2.2"
pytest-assume = "^2.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]
